refactor: CorrectionItemService 분리 및 에러 처리 계층 정리 (#44)
- CorrectionItemService 생성하여 CorrectionItemRepository 의존을 PortfolioCorrectionService에서 분리
- Repository의 NOT_FOUND throw를 Service 단으로 이동 (아키텍처 규칙 반영)
- PortfolioCorrectionRepository, PortfolioRepository, UserRepository에서 null 반환으로 변경
- ExternalPortfolioFacade에 CorrectionItemService 주입
