+++ src/modules/portfolio-correction/application/facades/external-portfolio.facade.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Transactional } from 'typeorm-transactional';
 import { ExternalPortfolioService } from 'src/modules/portfolio/application/services/external-portfolio.service';
 import { PortfolioCorrectionService } from '../services/portfolio-correction.service';
+import { CorrectionItemService } from '../services/correction-item.service';
 import { StructuredPortfolioResDto } from '../dtos/external-portfolio.dto';
 import { MAX_EXTERNAL_PORTFOLIO_BLOCKS } from 'src/modules/portfolio/domain/portfolio.entity';
 import { BusinessException } from 'src/common/exceptions/business.exception';
@@ -11,13 +12,14 @@ import { ErrorCode } from 'src/common/exceptions/error-code.enum';
 export class ExternalPortfolioFacade {
     constructor(
         private readonly externalPortfolioService: ExternalPortfolioService,
-        private readonly portfolioCorrectionService: PortfolioCorrectionService
+        private readonly portfolioCorrectionService: PortfolioCorrectionService,
+        private readonly correctionItemService: CorrectionItemService
     ) {}
 
     async getExternalPortfolios(correctionId: number): Promise<StructuredPortfolioResDto[]> {
         await this.portfolioCorrectionService.findByIdOrThrow(correctionId);
         const portfolioIds =
-            await this.portfolioCorrectionService.findPortfolioIdsByCorrectionId(correctionId);
+            await this.correctionItemService.findPortfolioIdsByCorrectionId(correctionId);
         const portfolios = await this.externalPortfolioService.getExternalPortfolios(portfolioIds);
         return portfolios.map((portfolio) => StructuredPortfolioResDto.from(portfolio));
     }
@@ -30,13 +32,13 @@ export class ExternalPortfolioFacade {
         const correction = await this.portfolioCorrectionService.findByIdOrThrow(correctionId);
 
         const currentCount =
-            await this.portfolioCorrectionService.countItemsByCorrectionId(correctionId);
+            await this.correctionItemService.countItemsByCorrectionId(correctionId);
         if (currentCount >= MAX_EXTERNAL_PORTFOLIO_BLOCKS) {
             throw new BusinessException(ErrorCode.CORRECTION_BLOCK_LIMIT_EXCEEDED);
         }
 
         const savedPortfolio = await this.externalPortfolioService.createEmptyPortfolio(userId);
-        await this.portfolioCorrectionService.createCorrectionItem(savedPortfolio, correction);
+        await this.correctionItemService.createCorrectionItem(savedPortfolio, correction);
 
         return StructuredPortfolioResDto.from(savedPortfolio);
     }
