+++ src/modules/portfolio-correction/application/services/correction-item.service.ts
@@ -1,0 +1,30 @@
+import { Injectable } from '@nestjs/common';
+import { CorrectionItemRepository } from '../../infrastructure/repositories/correction-item.repository';
+import { CorrectionItem } from '../../domain/correction-item.entity';
+import { Portfolio } from 'src/modules/portfolio/domain/portfolio.entity';
+import { PortfolioCorrection } from '../../domain/portfolio-correction.entity';
+
+@Injectable()
+export class CorrectionItemService {
+    constructor(private readonly correctionItemRepository: CorrectionItemRepository) {}
+
+    saveCorrectionItem(correctionItem: CorrectionItem): Promise<CorrectionItem> {
+        return this.correctionItemRepository.save(correctionItem);
+    }
+
+    async createCorrectionItem(
+        portfolio: Portfolio,
+        correction: PortfolioCorrection
+    ): Promise<CorrectionItem> {
+        const item = CorrectionItem.create(portfolio, correction);
+        return this.correctionItemRepository.save(item);
+    }
+
+    async findPortfolioIdsByCorrectionId(correctionId: number): Promise<number[]> {
+        return this.correctionItemRepository.findPortfolioIdsByCorrectionId(correctionId);
+    }
+
+    countItemsByCorrectionId(correctionId: number): Promise<number> {
+        return this.correctionItemRepository.countByCorrectionId(correctionId);
+    }
+}
