+++ src/modules/user/application/services/user.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { UserProfileResDto } from '../dtos/user-profile.dto';
 import { UserRepository } from '../../infrastructure/repositories/user.repository';
+import { BusinessException } from 'src/common/exceptions/business.exception';
+import { ErrorCode } from 'src/common/exceptions/error-code.enum';
 
 @Injectable()
 export class UserService {
@@ -8,6 +10,9 @@ export class UserService {
 
     async getProfile(userId: number): Promise<UserProfileResDto> {
         const profile = await this.userRepository.findByIdWithProfile(userId);
+        if (!profile) {
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND);
+        }
         return UserProfileResDto.from(profile);
     }
 }
